added sweet alert for delete  category

+++ Front End/src/components/Dashboard/Category/Category.js
@@ -9,7 +9,7 @@ import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import CategoryModal from './CategoryModal/CategoryModal';
 import NotFoundMessage from '../../NotFoundMessage/NotFoundMessage';
-
+import Alert from 'sweetalert2'
 
 
 
@@ -40,15 +40,35 @@ const Category = () => {
     }
 
     const deleteProduct = async(id)=>{
+        const {isConfirmed} = await   Alert.fire({
+            title: 'Are you sure want to delete?',
+            text: "You can add new category again!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes'
+          });
+    if(isConfirmed){
         const response = await fetch(`http://localhost:5000/api/v1/category/delete/${id}`,{
             method:"DELETE"
         })
         const {success, message } = await response.json();
         if(success){
-            toast.success(message,{position: "top-center",autoClose: 1000});
+            setLoading(false);
+            Alert.fire(
+                message,
+                'Category has been deleted!.',
+                'success'
+              ) 
         }else{
-            toast.error(message,{position: "top-center",autoClose: 1000});
+            Alert.fire({
+                icon: 'error',
+                title: message,
+              })
         }
+    }
+     
     }
